Add new packet_header_length_error exception.
Previously we had a single invalid_header_error for all issues with the
packet header, but this makes it difficult to distinguish between
malformed packets and those that are truncated, important when trying
to process streaming data.

Now, we have a packet_header_length_error that will be thrown when
a data buffer is shorter than required by a valid header.